library-authentication/src/main/java/ru/otus/library/service/library/impl/AuthorServiceImpl.java
package ru.otus.library.service.library.impl;

import org.springframework.stereotype.Service;
import ru.otus.library.data.library.model.Author;
import ru.otus.library.repository.AuthorRepository;
import ru.otus.library.service.library.AuthorService;

import java.util.Collections;
import java.util.List;
import java.util.Optional;

@Service
public class AuthorServiceImpl implements AuthorService {

    private final AuthorRepository authorRepository;

    public AuthorServiceImpl(AuthorRepository authorRepository) {
        this.authorRepository = authorRepository;
    }

    @Override
    public void saveAuthor(Author author) {
        authorRepository.save(author);
    }

    @Override
    public List<Author> findAuthorsByName(String name) {
        if (name == null || name.equals("")) {
            return Collections.EMPTY_LIST;
        }
        return authorRepository.findByName(name);
    }

    @Override
    public List<Author> findAllAuthors() {
        return authorRepository.findAll();
    }

    @Override
    public void deleteAuthor(Long id) {
        authorRepository.deleteById(id);
    }

    @Override
    public Optional<Author> getAuthorById(Long id) {
        return authorRepository.findById(id);
    }

    @Override
    public void updateAuthor(Author author) {
        authorRepository.save(author);
    }
}
